qemu-log: wrap long text
The existing help output is a bit hard to read due to the ad-hoc wrapping.
This makes it a bit more programmatic (at least, it wraps it once, but
that should be good enough for now).  It's not terribly intelligent (just
sticks a dash in the middle of the broken word) rather than splitting on
whitespace, but not sure how much time we really want to invest here.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>